Fix underflow in pieces_threatened_by_pawns
I am very upset with this sprt
Results of engine1 vs engine2 (8+0.08, NULL, 16MB, Pohl.epd):
Elo: -3.42 +/- 4.73, nElo: -5.34 +/- 7.38
LOS: 7.80 %, DrawRatio: 45.39 %, PairsRatio: 0.94
Games: 8522, Wins: 2359, Losses: 2443, Draws: 3720, Points: 4219.0 (49.51 %)
Ptnml(0-2): [200, 1002, 1934, 932, 193], WL/DD Ratio: 1.17
LLR: -0.24 (-10.7%) (-2.25, 2.89) [-10.00, 0.00]

Bench Nodes: 11120207 NPS: 11876282 Time: 936.3374ms

According to godbolt with `-C opt-level=3 -Ctarget-cpu=znver4` these all produce the exact same code:

```
\#[unsafe(no_mangle)]
fn test_fn(n1: u64, n2: u64) -> i16 {
    (n1.count_ones() - n2.count_ones()) as i16
}

\#[unsafe(no_mangle)]
fn test_fn2(n1: u64, n2: u64) -> i16 {
    n1.count_ones() as i16 - n2.count_ones() as i16
}

\#[unsafe(no_mangle)]
fn test_fn3(n1: u64, n2: u64) -> i16 {
    (n1.count_ones() as i32 - n2.count_ones() as i32) as i16
}
```

+++ src/evaluate.rs
@@ -87,7 +87,7 @@ impl Board {
                 / (MAX_GAME_STAGE - ENDGAME_GAME_STAGE_FOR_QUIESCENSE);
         }
 
-        let pieces_threatened_by_pawns = (self.get_pieces_threatened_by_pawns(true).count_ones() - self.get_pieces_threatened_by_pawns(false).count_ones()) as i16;
+        let pieces_threatened_by_pawns = self.get_pieces_threatened_by_pawns(true).count_ones() as i16 - self.get_pieces_threatened_by_pawns(false).count_ones() as i16;
         midgame_values += pieces_threatened_by_pawns * PIECES_THREATENED_BY_PAWNS.midgame;
         endgame_values += pieces_threatened_by_pawns * PIECES_THREATENED_BY_PAWNS.endgame;
